Make hostname and socket configurable
This allows us to run this in a non-nerves instances, just add `runtime: false` option to the deps line and add configuration.

```
config :pigpiox, hostname: 'somehost.local', port: 8888
```

And start all the children you will need, I suggest the socket and the watcher

To your application add the following children:
```
      {Pigpiox.Socket, [name: Pigpiox.Socket]},
      {Pigpiox.GPIO.WatcherSupervisor, []}
```

If config is left out it will try the default, localhost and 8888.